chore: move info modal to modals dir
