+++ frontend/src/components/NavBar.tsx
@@ -16,7 +16,7 @@ import MenuItem from '@mui/material/MenuItem'
 import { useAppDispatch } from '../store/hooks'
 import { setNav } from '../store/globalSlice'
 import logo from '../assets/images/team-logo-light.png'
-import { InfoModal } from './InfoModal'
+import { InfoModal } from './modals/InfoModal'
 import { useLocation, useNavigate } from 'react-router-dom'
 import { updateFilterAccordingToNav } from '../store/filterSlice'
 import { Topology } from '../enum'
