+++ frontend/src/components/modals/InfoModal.tsx
@@ -7,7 +7,7 @@ import { Box, Modal, IconButton, Button } from '@mui/material'
 import CloseIcon from '@mui/icons-material/Close'
 import ChevronRight from '@mui/icons-material/ChevronRight'
 import InfoRoundedIcon from '@mui/icons-material/InfoRounded'
-import logo from '../assets/images/team-logo-no-margins.png'
+import logo from '../../assets/images/team-logo-no-margins.png'
 import FilterListIcon from '@mui/icons-material/FilterList'
 import BarChartIcon from '@mui/icons-material/BarChart'
 import SearchIcon from '@mui/icons-material/Search'
